Fix missing RBAC rules
This is the first attempt to add required rbac rules to individual
operators so that operators can manipulate the resources. Note that
this is not yet tested completely so there might be some more rules
we have to add later, but we can fix apparently missing rules now to
(hopefully) make the later attempt easier.