Squash config files into a single layer
... instead of creating layer per service. This allows us to maintain
all options in a single heat.conf file instead we maintain the config
file for each service. So far we are not aware of any specific
requirement to tune options differently in each service by default.
Users can use custom config interface in Api/Engine CR in case they
need to tune a parameter for a specific service.