Fix selector in affinity policy
Currently we use the same name in the affinity policies to distribute
pods, but this is wrong and we don't have to allocate pod for different
components in different nodes. This fixes the selector and ensures we
use the different name for each component.

In addition, deployment name is now determined based on CR instance
name, otherwise two instances can create two deployments with the same
name,