Fix selector in affinity policy
Currently we use the same name in the affinity policies to distribute
pods, but this is wrong and we don't have to allocate pod for different
components in different nodes. This fixes the selector and ensures we
use the different name for each component. Because of the current
limitation in affinity utilities in lib-common this removes usage of
component selector and squashes a component name into the app selector.

This also updates name of deployments so that users can distinguish
what service is managed by a specific deployment more easily.