Update Heat kuttl sample to test webhook
By removing parameters that will receive their defaults from the
webhook, we can validate functionality in our kuttl tests. This change
updates the template being used to create the Heat deployment, based on
the current config/sample. It additionally removes the `containerImage`
parameter to assert whether the value is applied via the webhook.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>