Stop registering schemas for default objects
Currently we register schemas for corev1 and appsv1 but this is
incomplete and we need a few more schemas like batchv1. However we have
not noticed it because the resources available in k8s by default don't
need explicit schema registration.

Let's remove these instead of keeping the incomplete list.