Block changing DatabaseInstance
We currently have no mechanism to validate the new database will match
the existing database. This could ultimatly result in user objects being
orphaned and unrecoverable. We may need to implement a bypass for this
webhook in cases where we are doing adoptions and have independently
verified the contents of the databases. But without any kind of
validation, we should block users from inadvertently causing potentially
unrecoverable errors.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>