Tidy up condition handling
This change does some tidy up of how we handle conditions. It deduplicates the
code that determines if we're initializing or copying conditions and separates
logic into distinct functions. Rather than the list of conditions
belonging to the reconciler, conditions relevant to the API are moved
into the api types for each controller where the logic in the reconciler
will reference them.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>