Fix default cache config for no tls
When env is deployed with tls disabled, cache config was wrong as "dogpile.cache.memcached" backend requires inet[6] prefixes for memcache_servers setting while it was set without the prefix. with IPv4 issue not observed as "inet" is default prefix, this patch fixes it by setting [cache]/memcache_servers based on tls config.

closes OSPRH-12225

Signed-off-by: Fabricio Aguiar <fabricio.aguiar@gmail.com>