Make sure we reconcile heat-engine secrets
We trigger reconciles for each controller based on different events.
This can lead to a situation where Heat engine - which doesn't have
the same number of events that will trigger a reconcile, becomes out
of sync with the other controllers. For example, the configs which are
essentially the same but updated with https endpoints can be synced
by the HeatAPI and HeatCfnApi controllers first, before the
HeatEngine controller finally catches up. In this interval of time
the heat-engine process will fail since it's trying to speak to non
https endpoints instead of the correct https ones.

This change switches the configMapFn() Watches function to secretFn
and switches out instances where we were watching for ConfigMap changes
to be watching for secret changes. We don't use ConfigMaps anymore in
Heat Operator, so this change was left over from an oversight rather
than functionally necessary.

Jira: https://issues.redhat.com/browse/OSPCIX-795
Signed-off-by: Brendan Shephard <bshephar@redhat.com>